requirements.txt
pennylane-forest
qiskit
qiskit-aqua
numpy
scipy
jupyter
matplotlib
pennylane
pyquil
tqdm
pytorch
